// apps/git-client/src/generated/Changelist.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RepositoryId } from "./RepositoryId";

export type Changelist = { id: string, repositoryId: RepositoryId, name: string, paths: Array<string>, createdAtMs: number, updatedAtMs: number, };